Add .hlint.yaml.
This is the HLint configuration file generated by running:

`hlint . --default > .hlint.yaml`

Notice that we are ignoring one hint:

`ignore: {name: "Parse error"}`

However, the source code does not actually contain a Parse error. Both
`ghcid` and `stack ghci` can load the source code fine.

I'd like to get rid of this ignore, but it is not as dangerous as it
might seem since, if the source code does result in a Parse error, other
checks would fail loudly.